feat-added-delete-confirm-dialogbox

+++ src/components/CategoryTable.tsx
@@ -14,6 +14,7 @@ import {
 } from "@/service/category/useCategoryService";
 import type { Category } from "@/types/category";
 import { Spinner } from "./ui/spinner";
+import DeleteConfirmDialog from "./DeleteConfirmDialog";
 
 type Props = {
   onUpdate: (category: Category) => void;
@@ -21,26 +22,26 @@ type Props = {
 
 export default function CategoryTable({ onUpdate }: Props) {
   const { data: categories = [], isPending, isError } = UseGetAllCategories();
-
-  const { mutateAsync: deleteCategory } = UseDeleteCategory();
+  const { mutateAsync: deleteCategory, isPending: isDeleting } = UseDeleteCategory();
 
   const handleDelete = async (id: number) => {
     await deleteCategory({ id });
   };
 
- if (isPending)
-  return (
-    <p className="flex items-center justify-center">
-      <Spinner/>
-    </p>
-  );
+  if (isPending)
+    return (
+      <p className="flex items-center justify-center mt-16">
+        <Spinner />
+      </p>
+    );
 
-if (isError)
-  return (
-    <p className="text-center text-red-500">
-      Failed to load categories
-    </p>
-  );
+  if (isError)
+    return (
+      <p className="text-center text-red-500 mt-16">
+        Failed to load categories
+      </p>
+    );
+
   return (
     <Table>
       <TableCaption>List of all categories</TableCaption>
@@ -70,26 +71,28 @@ if (isError)
             </TableCell>
 
             <TableCell className="text-center">
-              <div className="flex justify-center gap-2">
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={() => onUpdate(cat)}
-                  className="border-blue-500 bg-blue-500 text-white hover:bg-blue-600 hover:text-white"
-                >
-                  Update
-                </Button>
+  <div className="flex justify-center gap-2">
+    {/* Update Button */}
+    <Button
+      
+      variant="outline"
+      onClick={() => onUpdate(cat)}
+      className="bg-blue-500 border-blue-500 text-white hover:bg-blue-600 hover:text-white"
+    >
+      Update
+    </Button>
 
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={() => handleDelete(cat.id)}
-                  className="border-red-500 bg-red-500 text-white hover:bg-red-600 hover:text-white"
-                >
-                  Delete
-                </Button>
-              </div>
-            </TableCell>
+    {/* Delete Button */}
+    <DeleteConfirmDialog
+      title="Delete category?"
+      itemName={cat.name}
+      onConfirm={() => handleDelete(cat.id)}
+      isLoading={isDeleting}
+      triggerLabel="Delete"
+      triggerClassName="bg-red-500 border-red-500 text-white hover:bg-red-600 disabled:opacity-50"
+    />
+  </div>
+</TableCell>
           </TableRow>
         ))}
       </TableBody>
